+++ db/migrate/20150906195818_add_locale_to_alchemy_languages.rb
@@ -1,0 +1,7 @@
+class AddLocaleToAlchemyLanguages < ActiveRecord::Migration
+  def change
+    add_column :alchemy_languages, :locale, :string
+    execute \
+      "UPDATE #{Alchemy::Language.table_name} SET locale = language_code WHERE locale IS NULL;"
+  end
+end