+++ spec/controllers/alchemy/messages_controller_spec.rb
@@ -181,7 +181,7 @@ module Alchemy
               end
 
               context "and mailer_config has no instructions for success_page" do
-                let(:language) { mock_model('Language', code: 'en', pages: double(find_by: build_stubbed(:alchemy_page))) }
+                let(:language) { mock_model('Language', code: 'en', locale: 'en', pages: double(find_by: build_stubbed(:alchemy_page))) }
 
                 before do
                   allow(controller).to receive(:mailer_config).and_return({})